% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coinertia.R
\name{check_coinertia}
\alias{check_coinertia}
\title{Test coinertia significance function}
\usage{
check_coinertia(R, L)
}
\arguments{
\item{R}{R table where rows are functions/pathways and columns are samples}

\item{L}{L table where rows are taxa/OTUs and columns are the same samples as table R}

\item{Requires}{the references functional (R) and its paired taxonomy table (L) and the taxonomy table (Q) for which coinertia and its significance will be calculated.}
}
\description{
This function will check for significance of coinertia between the paired taxonomy dataset and its functional profile dataset.
Ensure that the datasets are normalised/transformed before implementing this command.
It will utilise the dudi.pca, dudi.coa, coinertia and randtest commands from the ade4 library and return the results as an list.
}
\examples{
check_coinertia()
}
\keyword{coinertia}
\keyword{significance}
