% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functionality.R
\name{filter_functionality}
\alias{filter_functionality}
\title{Filters the functional dataset based on a average function/pathway coverage or abundance threshold}
\usage{
filter_functionality(dataset, type, threshold_dataset, threshold)
}
\arguments{
\item{dataset}{dataset where rows are functions/pathways and columns are samples.}

\item{type}{the type of dataset provided: (kegg or metacyc).}

\item{threshold_dataset}{dataset containing the information about pathway/function coverage. Threshold dataset should contain the same functions/pathways as present in dataset.}

\item{Requires}{the dataset to be transformed and whether to filter by covearge or abundance.}

\item{Optional}{threshold_dataset, levels.}

\item{levels}{Default is 0.01 for coverge or the first quartile of the mean function/pathway abundance.}
}
\description{
This function will filter the functional dataset and remove the low coverage/abundance functions/pathways.
By default, use the pathway coverage dataset for filtering. If coverage dataset is not available use the pathway abundance as threshold. 
You can opt to not filter the dataset at all so that all functionality is predicted.
}
\examples{
filter_functionality()
}
\keyword{filter}
\keyword{threshold}
