% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCO.R
\name{IPCO}
\alias{IPCO}
\title{IPCO function}
\usage{
IPCO(R, L, Q)
}
\arguments{
\item{R}{R dataset where rows are functions/pathways and columns are samples}

\item{L}{L dataset where rows are taxa/OTUs and columns are the same samples as dataset R}

\item{Q}{Q dataset where rows are taxa/OTUs common between L and Q and columns are the samples for which functionality will be inferred}

\item{Requires}{the references functional (R) and its paired taxonomy dataset (L) and the taxonomy dataset (Q) for which functions need to be inferred}
}
\description{
This function will implement the IPCO algorithm to generate an inferred functionality matrix for your taxonomy dataset.
Ensure that the datasets are normalised/transformed before implementing this command.
Check for significance of coinertia between R and L if using a different reference dataset than the one provided. If the coinertia between R and L is not significant, results maybe unreliable.
}
\examples{
IPCO()
}
\keyword{Inferring}
\keyword{functionality}
