% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCO_Healthy.R
\docType{data}
\name{IPCO_Healthy}
\alias{IPCO_Healthy}
\title{Healthy cohort}
\format{An object of class list}
\usage{
load_ref(IPCO_Healthy)
}
\description{
Healthy cohort
}
\examples{
load_ref()
}
\keyword{datasets}
