% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCO_HMP.R
\docType{data}
\name{IPCO_HMP}
\alias{IPCO_HMP}
\title{HMP data}
\format{An object of class list}
\usage{
load_ref(samples="HMP")
}
\description{
HMP data
}
\examples{
load_ref(samples="HMP")
}
\keyword{datasets}
