% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ref.R
\name{load_ref}
\alias{load_ref}
\title{Loads IPCO reference datasets}
\usage{
load_ref(samples)
}
\arguments{
\item{samples}{use samples="HMP" if you want to upload HMP dataset rather than using the cohort of healthy samples as reference}
}
\description{
This function will load the references datasets provided by IPCO as default
If you want to use any other datasets as reference, then this function is not required
}
\examples{
load_ref()
}
\keyword{data}
\keyword{reference}
