% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transforms the datasets using Hellinger transformations for IPCO implementation}
\usage{
transform_data(dataset, type)
}
\arguments{
\item{dataset}{dataset where rows are functions/pathways/OTUs and columns are samples}

\item{type}{Enter the type of your dataset values: proportion, counts or relab (relative abundance)}

\item{Requires}{the dataset to be transformed and the type of values present in it.}
}
\description{
This function will transforms the datasets using Hellinger transformations to be used for IPCO. It will return the dataset in matrix format.
If you want to use any other normalisation/transformation, then you can do it separately and use of this command is not required.
}
\examples{
transform_data()
}
\keyword{Hellinger}
\keyword{transformation}
